T0Auxiliares

Description

Módulo para o uso de auxiliares da parte Gráfica do jogo.

Desenhar Menu

drawMenu :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro horizontalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawMenu2 :: ([Tipos], [Picture]) -> Int -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas verticalmente.

Isto é, vai desenhando as colunas (verticalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

opcoes :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Desenhar Mapa

Modo BlockDude

drawMapaN :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro verticalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawMapa2N :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas horizontalmente.

Isto é, vai desenhando as linhas (horizontalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawPlayerN :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

Ideologia de todas as outras, no entanto aplica-se ao jogador, simples quanto isto!

pecasN :: (Peca, [Picture]) -> Float -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Modo DragonBall

drawMapaDB :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro verticalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawMapa2DB :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas horizontalmente.

Isto é, vai desenhando as linhas (horizontalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawPlayerDB :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

Ideologia de todas as outras, no entanto aplica-se ao jogador, simples quanto isto!

pecasDB :: (Peca, [Picture]) -> Float -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Modo AmongUs

drawMapaA :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro verticalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawMapa2A :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas horizontalmente.

Isto é, vai desenhando as linhas (horizontalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawPlayerA :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

Ideologia de todas as outras, no entanto aplica-se ao jogador, simples quanto isto!

pecasA :: (Peca, [Picture]) -> Float -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Desenhar Niveis

drawNiveis :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro horizontalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawNiveis2 :: ([Tipos], [Picture]) -> Int -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas verticalmente.

Isto é, vai desenhando as colunas (verticalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

levels :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Desenhar Instruções

drawInstrucoes :: ([Tipos], [Picture]) -> Int -> [Picture]

Muito semelhante às anteriores, no entanto, aqui só precisamos desenhar verticalmente e portanto não ha necessidade de auxiliares. Varia em y.

instruct :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Desenhar Vitória

drawVitoria :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

Através do Translate e do uso da lista de listas, é feito um distanciamento entre as Pictures. Desta forma, é desenhado primeiro horizontalmente, portanto, começamos por usar o Translate para que, a cada elemento de uma lista seja feito o distanciamento em x.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

drawVitoria2 :: ([Tipos], [Picture]) -> Int -> [Picture]

À semelhança da função anterior, aqui também é feito um distanciamento de Pictures, mas verticalmente.

Isto é, vai desenhando as colunas (verticalmente), usando também o Translate mas em y.

Os inteiros em Translate variam de acordo com aquilo que for o objetivo de reproduzir na tela.

victory :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Desenhar Background

backgroundsN :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

backgroundsDB :: (Tipos, [Picture]) -> Picture

Estas funções são o que fazem variar o tema também, portanto, criamos uma para cada caso para depois ir alterando de acordo com cada estado.

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

backgroundsA :: (Tipos, [Picture]) -> Picture

Esta função utiliza os elementos da lista para atribuir correspondência a cada Imagem desejada!

Varia também de acordo com aquilo que cada um pretende reproduzir.

Detetar Porta

porta :: Jogo -> (Int, Int)

Esta é a função que obtém as coordenadas da Porta para poder, durante a reageEvento avaliar se é o fim do nível ou não. Utiliza findPorta como auxiliar.

findPorta :: Mapa -> Int -> Int -> (Int, Int)

findPorta irá ler as listas até encontrar a porta e assim devolver as coordenadas corretas. Isto é, se a cada elemnto de uma lista não encontrar, irá adicionar 1 a x, se no fim da lista não tiver encontrado e for para a seguinte, então aumenta 1 a y.