T0Dados

Description

Tipos de dados e funções auxiliares para a realização do projeto de LI1 em 2021/22.

Tipos de dados

Básicos

type Coordenadas = (Int, Int)

Par de coordenadas de uma posição no Mapa.

data Direcao

Direção de um Jogador no Mapa.

Constructors

Este

Oeste

Mapas

type Mapa = [[Peca]]

data Peca

Uma peça no Mapa.

Constructors

Bloco
um bloco que é indestrutível e não movivel

Caixa
a caixa é como um bloco mas pode ser movida pelo Jogador

Porta
a porta é a posição final do jogo

Vazio

PersonagemD

PersonagemE

BackgroundGame

Jogo

data Jogo

O nível de um jogo, que inclui o puzzle (mapa) e o personagem (jogador).

Constructors

Jogo

Fields

Mapa

o puzzle em si

Jogador

o personagem do jogo

data Jogador

O personagem que é controlado pelo Jogador.

Constructors

Jogador

Fields

Coordenadas

a posição atual no Mapa

Direcao

a direção atual

Bool

um booleano que indica se o Jogador está a carregar uma Caixa ou não

data Movimento

Os movimentos que podem ser tomados pelo jogador em cada estado do Jogo.

Constructors

AndarEsquerda
a acção de andar para a esquerda

AndarDireita
a ação de andar para a direita

Trepar
a ação de trepar uma caixa ou bloco

InterageCaixa
a ação de pegar ou largar uma caixa

data Tipos

Tipos representa um vasto tipo de informações, já que o mesmo podia ser utilizado em várias funções, decici utilizar uma só 'data';

Constructors

Nada

Select

Play

Level

Tips

Exit

Back

Lvl1

Lvl2

Lvl3

Lvl4

Lvl5

Lvl6

Lvl7

Lvl8

Lvl9

Lvl10

Texto

WinCima

WinBaixo

Continue

NiveisGeral

OpcoesGeral

GameGeral

BlockDude

DragonBall

AmongUs

Theme

Show1

Show2

Show3

Show4

Show5

Show6

Show7

Show8

Show9

Show10

Theme1

Theme2

Theme3

ATheme

DBTheme

NTheme

data Jogos

Jogos é utilizado como acumulador ou método de "mémoria" pois é essa 'data' que organiza o nível que se está;

Constructors

Jogo1

Jogo2

Jogo3

Jogo4

Jogo5

Jogo6

Jogo7

Jogo8

Jogo9

Jogo10

data OpcoesPrincipal

OpcoesPrincipal representa as opções que irão aparecer no menu inicial;

Constructors

Jogar

Niveis

Continuar

Instrucoes

Tema

Sair

data OpcoesNiveis

OpcoesNiveis semelhante ao de cima mas para os niveis;

Constructors

Nivel1

Nivel2

Nivel3

Nivel4

Nivel5

Nivel6

Nivel7

Nivel8

Nivel9

Nivel10

Voltar

data Menu

Menu é o que vai definir todas as anteriores de modo a poder dar base ao jogo.

Constructors

Controlador OpcoesPrincipal Tipos

Selecionador OpcoesNiveis Tipos

ModoJogo Jogo Jogo Jogos Tipos

Pergaminho Tipos

ImagemVitoria OpcoesPrincipal Jogo Jogos Tipos

Temas Tipos Tipos