T1ValidaMapa

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

Validar Blocos

validablocos :: [(Peca, Coordenadas)] -> Bool

A função validablocos retorna um Bool se o mapa não tiver pecas nas mesmas coordenadas. Sendo composta por 5 auxiliares da seguinte forma:

>>> validablocos [(Bloco,(5,5)),(Caixa,(5,5))]
False
>>> validablocos [(Bloco,(0,1)),(Bloco,(1,1))]
True

Auxiliares

validablocosaux :: [(Peca, Coordenadas)] -> [Coordenadas]

Função usada para simplificar a lista devolvendo apenas uma lista com as Coordenadas

>>> validablocosaux [(Bloco,(5,2)),(Caixa,(3,3))]
[(5,2),(3,3)]

mergelists :: [Coordenadas] -> [[Coordenadas]]

Faz um group na lista composta apenas por coordenadas de modo a juntar numa só lista.

>>> mergelists [(2,1),(5,1),(2,1)]
[[(2,1),(2,1)],[5,1]]

testbool :: [[Coordenadas]] -> [Coordenadas] -> Bool

Ligada à função anterior testbool, mede o tamanho da lista obtida por mergelist e o da obtida por validablocosaux. Em caso de resultado diferente, devolve um False

>>> testbool [[(2,1)],[(3,1),(3,1)]] [(3,1),(2,1),(3,1)]
False

Validar Porta

validaporta :: [(Peca, Coordenadas)] -> Bool

Composta por uma auxiliar, validaporta permite concluir se existe uma e uma só Porta no mapa.

>>> [(Bloco,(3,2)),(Porta,(2,2))]
True

Auxiliar

portaaux :: [(Peca, Coordenadas)] -> [Peca]

portaaux devolve uma lista em que as peças são iguais a Porta.

>>> portaaux [(Bloco,(2,1))]
[ ]

>>> portaaux [(Bloco,(2,1)),(Porta,(2,2))]
[Porta] 

Validar Caixas

validaCaixaFlutua :: [(Peca, Coordenadas)] -> Bool

Função que avalia se nenhuma Caixa flutua, ou seja, se todas as peças Caixa têm por baixo outra Caixa ou um Bloco e se a ultima Caixa está em cima de um Bloco.

>>> validaCaixaFlutua [(Bloco,(1,2)),(Caixa,(1,1)),(Bloco,(2,3)),(Vazio,(4,5)),(Caixa,(2,2)),(Caixa,(2,1)),(Vazio,(0,0)),(Bloco,(0,2))]
True

>>> validaCaixaFlutua [(Vazio,(0,0)),(Vazio,(0,1)),(Bloco,(0,2)),(Bloco,(1,2)),(Bloco,(2,2)),(Caixa,(2,1)),(Vazio,(3,2)),(Caixa,(3,1))]
False

Auxiliares

validaCaixa :: [Coordenadas] -> [Coordenadas] -> [Coordenadas] -> Bool

Função que recebe a lista de coordendas das peças Caixa, a lista de coordenadas de Caixa, subtraindo uma unidade a y, e a lista de coordenadas das peças Bloco subtraindo uma unidade a y. A partir destas listas, testa se nenhuma Caixa flutua.

>>> validaCaixa [(1,0),(2,1),(2,0)] [(1,0),(2,1),(2,0)] [(1,1),(2,2),(0,1)]
True

>>> validaCaixa [(2,1),(3,1)] [(2,0),(3,0)] [(0,1),(1,1),(2,1)]
False

armazenaCaixaCoord :: [(Peca, Coordenadas)] -> [Coordenadas]

Função que armazena as coordenadas de todas as peças Caixa do mapa.

>>> armazenaCaixaCoord [(Bloco,(1,2)),(Caixa,(1,1)),(Bloco,(2,3)),(Vazio,(4,5)),(Caixa,(2,2)),(Caixa,(2,1)),(Vazio,(0,0)),(Bloco,(0,2))]
[(1,1),(2,2),(2,1)]

>>> armazenaCaixaCoord [(Vazio,(0,0)),(Vazio,(0,1)),(Bloco,(0,2)),(Bloco,(1,2)),(Bloco,(2,2)),(Caixa,(2,1)),(Vazio,(3,2)),(Caixa,(3,1))]
[(2,1),(3,1)]

armazenaCaixaCoord1 :: [(Peca, Coordenadas)] -> [Coordenadas]

Função que armazena as coordenadas de toda a Caixa do mapa, subtraindo-lhes 1 unidade a y.

>>> armazenaCaixaCoord1 [(Bloco,(1,2)),(Caixa,(1,1)),(Bloco,(2,3)),(Vazio,(4,5)),(Caixa,(2,2)),(Caixa,(2,1)),(Vazio,(0,0)),(Bloco,(0,2))]
[(1,0),(2,1),(2,0)]

>>> armazenaCaixaCoord1 [(Vazio,(0,0)),(Vazio,(0,1)),(Bloco,(0,2)),(Bloco,(1,2)),(Bloco,(2,2)),(Caixa,(2,1)),(Vazio,(3,2)),(Caixa,(3,1))]
[(2,0),(3,0)]

armazenaBlocoCoord :: [(Peca, Coordenadas)] -> [Coordenadas]

Função que armazena as coordenadas de todo o Bloco do mapa, subtraindo-lhes 1 unidade a y.

>>> armazenaBlocoCoord [(Bloco,(1,2)),(Caixa,(1,1)),(Bloco,(2,3)),(Vazio,(4,5)),(Caixa,(2,2)),(Caixa,(2,1)),(Vazio,(0,0)),(Bloco,(0,2))]
[(1,1),(2,2),(0,1)]

>>> armazenaBlocoCoord [(Vazio,(0,0)),(Vazio,(0,1)),(Bloco,(0,2)),(Bloco,(1,2)),(Bloco,(2,2)),(Caixa,(2,1)),(Vazio,(3,2)),(Caixa,(3,1))]
[(0,1),(1,1),(2,1)]

Validar Vazio

validavazio :: [(Peca, Coordenadas)] -> Bool

Esta função, composta por 3 auxiliares, testa se existe pelo menos um Vazio no mapa. Caso exista um Vazio definido no mapa, a função devolve logo um True. Caso contrário, a função irá testar se existe Vazio por omissão.

>>> validavazio [(Bloco,(0,0)),(Bloco,(1,0)),(Bloco,(2,1))]
True

Auxiliares

validavazioaux :: [(Peca, Coordenadas)] -> Int -> Int

Esta função irá contar o número de Vazio fornecidos ao mapa.

>>> validavazioaux [(Bloco,(5,9)),(Caixa,(10,5)),(Vazio,(4,2)),(Vazio,(3,3)),(Bloco,(2,9))] 0
2

medidas :: [(Peca, Coordenadas)] -> (Int, Int)

A função medidas mede o tamanho do mapa em x e em y para auxiliar a funçao seguinte e permitir descobrir se existe Vazio declarados por omissão.

>>> medidas [(Bloco,(5,9)),(Caixa,(10,5)),(Vazio,(4,2)),(Vazio,(3,3)),(Bloco,(2,9))]
(11,10)

omitidos :: (Int, Int) -> [(Peca, Coordenadas)] -> Bool

omitidos calcula a área do mapa com a função medidas anterior e compara ao número de peças declaradas. Caso a diferença seja diferente de 0, omitidos devolve um True .

validabase

validabase :: [(Peca, Coordenadas)] -> Bool

Esta função mais complexa, constituida por 7 funções auxiliares, testa a base do mapa, não podendo haver espaços vazios, impedindo quedas.

>>> validabase [(Bloco,(0,0)),(Bloco,(1,0)),(Bloco,(2,1)),(Bloco(3,0))]
True

>>> validabase [(Bloco,(0,0)),(Bloco,(2,0)),(Bloco,(2,1)),(Bloco(3,0))]
False

Auxiliares

justBlocks :: [(Peca, Coordenadas)] -> [Coordenadas]

justBlocks, tal como o nome indica, simplifica a lista, devolvendo apenas uma lista com as coordenadas das peças Bloco.

justxBlocks :: [Coordenadas] -> [Int]

justxBlocks simplifica a função simplificada, de modo a devolver uma lista apenas com os x das peças Bloco.

blocksbycolumn :: [Coordenadas] -> Bool

A função blocksbycolumn permite testar se em cada coluna, existe pelo menos um Bloco devolvendo um Bool.

superGroup :: (Eq a, Ord a) => [(a, b)] -> [(a, [b])]

superGroup é funçao muito especifica que converte os pares com x igual, num unico par com x, e com os y concatenados numa só lista.

>>> superGroup [(0,1),(1,3),(0,2)]
[(0,[1,2]),(1,[3])]

superGroupSort :: [(Int, [Int])] -> [(Int, [Int])]

superGroupSort é uma funçao que ordena a lista dentro do par, ou seja, ordena os y por ordem decrescente.

>>> superGroupSort [(1,[6,2,4,8,5]),(3,[51,2,4,5,4])]
[(1,[8,6,5,4,2]),(3,[51,5,4,4,2])]

order :: [(Int, [Int])] -> [Coordenadas]

Esta funçao devolve as coordenadas com maior y de cada coluna, ou seja, de cada lista dada no segundo elemento de cada par. No entanto, caso haja um muro ou queda, esta função irá devolver mais que um y pois se trata de uma grande parede que não pode ter buraco.

>>> order [(0,[9,7,1]),(1,[7,6,5,4])]
[(0,9),(1,7),(1,6),(1,5)]

mainfunction :: [(Int, [Int])] -> Bool

Esta é a função principal, pois irá testar se existe uma ligação em toda a base, ou seja, não há espaços vazios lá. Para isso, irá ser sempre testado, para todos os casos possíveis, se o y de x está na lista dos y para x+1, ou seja, se no x a seguir há alguma continuidade de chão. Caso haja, devolve um True, caso contrário, devolve um False!

>>> mainfunction [(0,[4])(1,[4,5,6]),(2,[7])]
True

>>> mainfunction [(0,[4])(1,[4,5,6]),(2,[8])]
False

testeauxiliar :: [Int] -> [Int] -> Bool

Esta função testa se um certo y em x se encontra na lista de y quando x+1 para saber então se existe continuação de mapa!