T3RepresentacaoJogo

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2021/22.

Documentation

show1 :: Jogo -> String

A partir de um jogador e de um mapa, com as pecas já organizadas por posições e linhas, retribuí uma string que é a representação textual do mapa.

Exemplos:

 Exemplo 1 : show1 (Jogo ([[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,2) Este False)) = "X   XnX   XnX> CXnXXXXX"

 Exemplo 2 : show1 (Jogo ([[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,1) Oeste True)) = "XC XnX<CXnXXXX"

separarLinhas :: Jogo -> Int -> String

Divide o mapa, que se encontra em string e já tem o jogador, em várias linhas, em que o número de linhas é igual à altura do mapa e o número de caractéres por linha é igual ao comprimento do mapa. Nesta função o acumulaador é responável por eliminar os caractéres que já estão nas linhas anteriores.

Exemplos:

 Exemplo 1 : separarLinhas (Jogo ([[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,2) Este False)) 0 = "X   XnX   XnX> CXnXXXXX"

 Exemplo 2 : separarLinhas (Jogo ([[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,1) Oeste True)) 0 = "XC XnX<CXnXXXX"

placeCaixa :: Jogo -> String

A partir de um mapa com o jogador já em string, que resultou da função auxiliar placeJogador, retira um espaço vazio (" ") no bloco por cima do jogador e coloca um C na mesma posição do espaço vazio caso o jogador carregue uma caixa. Caso o jogador não carregue uma caixa, a função devolve apenas a string dada pela função auxiliar.

Exemplos:

 Exemplo 2 : placeCaixa (Jogo ([[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,1) Oeste True)) = "XC XX<CXXXXX"

placeJogador :: Jogo -> String

A partir de um mapa já em string, que resultou da função auxiliar placePecas, retira um espaço (" "), que representa um vazio, e coloca um ou, que representa o jogador, na mesma posição, conforme o jogador estiver virado para Oeste ou Este.

Exemplos:

 Exemplo 1 : placeJogador (Jogo ([[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,2) Este False)) = "X   XX   XX> CXXXXXX"

 Exemplo 2 : placeJogador (Jogo ([[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,1) Oeste True)) = "X  XX<CXXXXX"

placePecas :: Jogo -> String

Transforma a lista de peças (Mapa) em uma string em que cada peça é representada por um caracter específico, sendo o Bloco representado por um X, a Caixa por um C, a porta por um P e o espaço vazio por um espaço (" ").

Exemplos:

 Exemplo 1 : placePecas (Jogo ([[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]) (Jogador (1,2) Este False)) = "X   XX   XX  CXXXXXX"

areaMapa :: Mapa -> Int

Calcula a área do mapa a partir da altura do mapa e da função auxiliar comprimentoMapa.

Exemplos:

 Exemplo 1 : areaMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] = 20

 Exemplo 2 : areaMapa [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco]] = 12

comprimentoMapa :: Mapa -> Int

Calcula o comprimento do mapa.

Exemplos:

 Exemplo 1 : comprimentoMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] = 5

Orphan instances

Show Jogo

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS