Tarefa2_2021li1g004

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

ConstroiMapa

Expand

Auxiliares

buildMatrix :: [(Peca, Coordenadas)] -> [[Peca]]

buildMatrix [] = []
buildMatrix [(peca,(x,y))] = replicate (y+1) (replicate (x+1) Vazio)
buildMatrix f@((peca,(x,y)):t) = replicate   
    ((maximum (listy f))+1) (replicate ((maximum (listx f))+1) Vazio)
        where 
            listx [] = []
            listx ((peca,(x,y)):t) = x:listx t    
            listy [] = []
            listy ((peca,(x,y)):t) = y:listy t

Função que cria uma matriz com as dimensões do mapa em que todos os elementos da mesma são Vazio. Ficando ja assim declarados também os elementos por omissão.

updateMatrix :: [(Peca, Coordenadas)] -> [[Peca]] -> [[Peca]]

Esta função substitui na matriz construida na auxiliar anterior, os espaços onde as peças são diferentes de Vazio. Ficando assim o Mapa do jogo concluido.

>>> updateMatrix [(Bloco,(1,1))] [[Vazio,Vazio],[Vazio,Vazio]]
[[Vazio,Vazio],[Vazio,Bloco]]

DesconstroiMapa

Expand

Auxiliares

tamanho :: [[Peca]] -> Int

Esta função auxilia na contagem de coordenadas, pois a cada vez que acaba a linha de x, adiciona +1 ao y.

cortavazio :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Elimina os blocos Vazio existentes na lista final de modo a ficarem apenas os importantes.

ordena :: [(Peca, Coordenadas)] -> [(Peca, Coordenadas)]

Tal como o nome indica, esta funçao ordena as peças para a desconstrução do mapa ser mais perceptivel.

function :: [[Peca]] -> [Peca] -> Coordenadas -> [(Peca, Coordenadas)]

Esta é a função que trata de obter as coordenadas para cada Peca. Isto é, calcula-se quantas listas tem na matriz, e sabe-se de quantos em quantos x se tem de aumentar o y. Assim, é possível calcular as coordenadas de cada bloco não Vazio.