Tarefa4_2021li1g004

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

correrMovimentos :: Jogo -> [Movimento] -> Jogo

É a generalização da função correrMovimentos.

moveJogador :: Jogo -> Movimento -> Jogo

Esta função aplica o efeito de um comando (i.e. Movimento) sobre o jogador.

Andar para a esquerda:

Caso o movimento introduzido seja AndarEsquerda, a função vai avaliar se o jogador pode realizar esse movimento através de uma série de condições. Condições estas:

1 - Existe um bloco à esquerda do jogador;

2 - Se o jogador carrega uma caixa, existe um bloco diretamente por cima do bloco que fica à esquerda do jogador.

Desta forma, caso o jogador não carregue uma caixa, este só alterará a sua posição se a condição 1 for falsa.

Por outro lado, se carregar uma caixa, só poderá alterar a sua posição se a condição 1 e 2 forem falsas.

Caso contrário a função irá devolver o mesmo Jogo, com o jogador virado para o lado que tentou andar mas não consegiu.

Andar para a direita:

Caso o movimento introduzido seja AndarDireita, a função vai avaliar se o jogador pode realizar esse movimento através de uma série de condições. Condições estas:

1 - Existe um bloco à direita do jogador;

2 - Se o jogador carrega uma caixa, existe um bloco diretamente por cima do bloco que fica à direita do jogador.

Desta forma, caso o jogador não carregue uma caixa, este só alterará a sua posição se a condição 1 for falsa.

Por outro lado, se carregar uma caixa, só poderá alterar a sua posição se a condição 1 e 2 forem falsas.

Caso contrário a função irá devolver o mesmo Jogo, com o jogador virado para o lado que tentou andar mas não consegiu.

Trepar:

Caso o movimento introduzido seja Trepar, a função vai avaliar se o jogador pode realizar esse movimento através de uma série de condições. Condições estas:

1 - O bloco que quer trepar é vazio;

2 - Existe um blococaixa por cima do blococaixa que o jogador quer trepar;

3 - O jogador carrega uma caixa:

Exite um bloco duas unidades a cima do bloco/caixa que o jogador quer trepar;

O jogador está virado para oeste.

4 - O jogador está virado para oeste.

Desta forma, caso o jogador não carregue uma caixa, este só poderá trepar se as condições 1 e 2 forem falsas para a direção em que este está virado.

Por outro lado, se carregar uma caixa, este só poderá trepar se as condições 1, 2 e 3 forem falsas para a direção em que o jogador está virado.

Caso contrário, a função irá devolver o mesmo Jogo.

Interagir com caixas:

Caso o movimento introduzido seja InterageCaixa, a função vai avaliar se o jogador pode realizar esse movimento através de uma série de condições. Condições estas:

1 - O jogador carrega uma caixa:

O jogador está virado para oeste;

O bloco à frente do jogador é vazio e o bloco diretamente por cima desse também o é;

Existe um bloco à frente do jogador, mas o bloco diretamente por cima é vazio.

2 - Existe um bloco por cima do jogador;

3 - Não existe uma caixa no bloco diretamente à frente do jogador;

4 - Existe um bloco/caixa por cima da caixa que o jogador quer pegar;

5 - O jogador está virado para oeste.

Desta forma, caso o jogador não carregue uma caixa, este só poderá pegar numa caixa se as condições 2, 3 e 4 forem falsas para a direção em que este está virado.

Por outro lado, se carregar uma caixa, este só a poderá largar se a condição 1.2 for verdadeira para a direção em que este está virado ou se a condição 1.3 for verdadeira para a direção em que este está virado.

Caso contrário, a função irá devolver o mesmo Jogo.

encontra :: Mapa -> Int -> Int -> Peca

Devolve um elemento de uma lista de listas.

Exemplos:

 Exemplo 1 : encontra [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] 2 1 = Porta

 Exemplo 2 : encontra [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] 1 1 = Vazio

encontra1 :: [[Peca]] -> Int -> [Peca]

Devolve uma lista de uma lista de listas.

Exemplos:

 Exemplo 1 : encontra1 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] 2 = [[Bloco,Porta,Vazio,Caixa,Bloco]]

 Exemplo 2 : encontra1 [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] 1 = [Bloco,Vazio,Vazio,Bloco]

encontra2 :: [Peca] -> Int -> Peca

Devolve um elemento de uma lista.

Exemplos:

 Exemplo 1 : encontra2 [Bloco,Porta,Vazio,Caixa,Bloco] 1 = Porta

 Exemplo 2 : encontra2 [Bloco,Vazio,Vazio,Bloco] 1 = Vazio

blocoCaixaEsquerdaSuperior :: Mapa -> Coordenadas -> Int

Calcula a ordenada do bloco ou caixa com menor ordenada à esquerda do jogador. Esta função só tem em conta as ordenadas dos blocos e caixas que têm ordenada inferior à do jogador.

Exemplos:

 Exemplo 1 : blocoCaixaEsquerdaSuperior [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (3,1) = 3

 Exemplo 2 : blocoCaixaEsquerdaSuperior [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] (2,2) = 3

blocosCaixasEsquerda :: Mapa -> Coordenadas -> Int -> [Int]

Calcula as ordenadas de todas caixas e blocos com abcissa igual à abcissa do jogador menos uma unidade e com ordenada superior à do jogador.

Exemplos:

 Exemplo 1 : blocosCaixasEsquerda [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (3,1) = [3]

 Exemplo 2 : blocosCaixasEsquerda [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] (2,2) = [3]

blocoCaixaDireitaSuperior :: Mapa -> Coordenadas -> Int

Calcula a ordenada do bloco ou caixa com menor ordenada à esquerda do jogador. Esta função só tem em conta as ordenadas dos blocos e caixas que têm ordenada inferior à do jogador.

Exemplos:

 Exemplo 1 : blocoCaixaDireitaSuperior [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Caixa,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = 2

 Exemplo 2 : blocoCaixaDireitaSuperior [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] (1,1) = 3

blocosCaixasDireita :: Mapa -> Coordenadas -> Int -> [Int]

Calcula as ordenadas de todas caixas e blocos com abcissa igual à abcissa do jogador mais uma unidade e com ordenada superior à do jogador.

Exemplos:

 Exemplo 1 : blocosCaixasDireita [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Porta,Caixa,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [2,3]

 Exemplo 2 : blocosCaixasDireita [[Bloco,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco]] (1,1) = [3]

adicionaCaixaOesteMapa1 :: Mapa -> Coordenadas -> Mapa

Substitui um espaço vazio por uma caixa quando o jogador larga uma caixa. Usa as funções auxiliares blocoCaixaEsquerdaSuperior e adicionaCaixaLinha. Esta função só se aplica se o bloco à frente do jogador for vazio.

Exemplos:

 Exemplo 1 : adicionaCaixaOesteMapa1 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : adicionaCaixaOesteMapa1 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (3,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Caixa,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

adicionaCaixaOesteMapa2 :: Mapa -> Coordenadas -> Mapa

Substitui um espaço vazio por uma caixa quando o jogador larga uma caixa. Usa a função auxiliar adicionaCaixaLinha. Esta função só se aplica se no bloco à frente do jogador existir um bloco ou uma caixa e o bloco por cima desse for vazio.

Exemplos:

 Exemplo 1 : adicionaCaixaOesteMapa2 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Bloco,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : adicionaCaixaOesteMapa2 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

adicionaCaixaEsteMapa1 :: Mapa -> Coordenadas -> Mapa

Substitui um espaço vazio por uma caixa quando o jogador larga uma caixa. Usa as funções auxiliares blocoCaixaEsquerdaSuperior e adicionaCaixaLinha. Esta função só se aplica se o bloco à frente do jogador for vazio.

Exemplos:

 Exemplo 1 : adicionaCaixaEsteMapa1 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : adicionaCaixaEsteMapa1 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

adicionaCaixaEsteMapa2 :: Mapa -> Coordenadas -> Mapa

Substitui um espaço vazio por uma caixa quando o jogador larga uma caixa. Usa a função auxiliar adicionaCaixaLinha. Esta função só se aplica se no bloco à frente do jogador existir um bloco ou uma caixa e o bloco por cima desse for vazio.

Exemplos:

 Exemplo 1 : adicionaCaixaEsteMapa2 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,vazio,Vazio,Bloco,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Vazio,Vazio,Bloco,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : adicionaCaixaEsteMapa2 [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

adicionaCaixaLinha :: [Peca] -> Coordenadas -> [Peca]

Substitui um espaço vazio (" ") por uma caixa numa linha do mapa.

Exemplos:

 Exemplo 1 : adicionaCaixaLinha Bloco,Vazio,Vazio,Vazio,Bloco = [Bloco,Vazio,Vazio,Caixa,Bloco]

 Exemplo 2 : adicionaCaixaLinha Bloco,Caixa,Vazio,Vazio,Bloco = [Bloco,Caixa,Vazio,Caixa,Bloco]

removeCaixaOesteMapa :: Mapa -> Coordenadas -> Mapa

Substitui um bloco com uma caixa por um bloco vazio. Esta função usa a função auxiliar removeCaixaLinha.

Exemplos:

 Exemplo 1 : removeCaixaOesteMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : removeCaixaOesteMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Caixa,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

removeCaixaEsteMapa :: Mapa -> Coordenadas -> Mapa

Substitui um bloco com uma caixa por um bloco vazio. Esta função usa a função auxiliar removeCaixaLinha.

Exemplos:

 Exemplo 1 : removeCaixaEsteMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,1) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

 Exemplo 2 : removeCaixaEsteMapa [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Caixa,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]] (2,2) = [[Bloco,Vazio,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Caixa,Vazio,Vazio,Bloco],[Bloco,Bloco,Bloco,Bloco,Bloco]]

removeCaixaLinha :: [Peca] -> Coordenadas -> [Peca]

Substitui uma caixa por um espaço vazio (" ") numa linha do mapa.

Exemplos:

 Exemplo 1 : removeCaixaLinha Bloco,Vazio,Vazio,Caixa,Bloco = [Bloco,Vazio,Vazio,Vazio,Bloco]

 Exemplo 2 : removeCaixaLinha Bloco,Caixa,Vazio,Caixa,Bloco = [Bloco,Caixa,Vazio,Vazio,Bloco]