Tarefa5_2021li1g083

Description

Tarefa 5 do projeto de LI1 em 2021/22

Este módulo, o principal, referente à tarefa 5, contém definições Haskell relativas à representação do Jogo

Documentation

data Menu

Menus do Jogo.

Constructors

MainMenu
Menu principal

GameMenu
Menu para iniciar novo jogo

MapMenu
Menu de seleção dos mapas

BetweenLevels
Menu entre níveis

Play
Opção em BetweenLevels que permite jogar o nivel selecionado.

MapEditor
Menu que permite ao jogador construir um mapa personalizado (sub-tarefa incompleta)

CreditsMenu
Menu que apresenta os Créditos do Jogo

data Options

Opções dentro dos menus.

Constructors

Mapa1
Nível 1

Mapa2
Nível 2

Mapa3
Nível 3

Mapa4
Nível 4

MapaB
Nível Bónus que só é possível acessar passando o nível 4, "Easter egg".

mainDisplay :: Display

Definições da janela do jogo

displayDimension :: (Int, Int)

"displayDimension" atribui uma dimensão à janela do jogo

type InGame = (Mapa, Jogador)

type Estado = Menu

type EstadoGloss = (Estado, Options, InGame, [Picture], [Picture], Peca)

estadoGlossInicial :: [Picture] -> [Picture] -> EstadoGloss

"estadoGlossInicial" define o estado inicial do gloss

doMovimento :: InGame -> Movimento -> Jogo

"doMovimento" executa movimentos no Jogo

jogoToEstado :: Jogo -> InGame

"jogoToEstado" troca o tipo do Jogo para InGame

reageEventoGloss :: Event -> EstadoGloss -> EstadoGloss

"reageEventoGloss" é a função que executa ações no jogo com base no menu em que o jogador se encontra e com base na opção escolhida

switchNiveis :: EstadoGloss -> Bool -> EstadoGloss

"switchNiveis" executa o mapa seguinte se o jogador concluir o nível ou dá reset/play again se o jogador quiser

reageTempoGloss :: Float -> EstadoGloss -> EstadoGloss

"reagetempoGloss" O intuíto da função era adicionar alguma ação ao passar do tempo mas acabamos por não o fazer, então retorna exatamente o estado anterior

desenhaEstadoGloss :: EstadoGloss -> Picture

"desenhaEstadoGloss" atribui uma imagem gráfica a cada menu definido ns struct Menu

fps :: Int

Definição dos frames por segundo, isto é, o número de vezes que a função reageTempoGloss é executada num segundo

main :: IO ()

A função "main" importa as imagens e corre o jogo.

sc :: Mapa -> Picture -> Picture

"sc" redimensiona os elementos do tipo Picture para a escala do jogo

gameScale :: Mapa -> Float

"gameScale"define a escala do jogo

pieceSide :: Mapa -> Float

"pieceSide" define o tamanho que cada Peca deve ter na janela do jogo

desenhaMapa :: Mapa -> Jogador -> Int -> Mapa -> [Picture] -> [Picture]

"desenhaMapa" desenha o jogo

desenhaLinha :: [Peca] -> Coordenadas -> Jogador -> Mapa -> [Picture] -> [Picture]

"desenhaLinha" é auxiliar da desenhaMapa e desenha cada linha de um Mapa na janela do gloss.