Main

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2021/22.

Type

type EstadoGloss = (Menu, (Float, Float), [Picture], [Picture], [Picture], [Picture], [Picture], [Picture], Float)

Nesta 'type':

Menu vai representar em que parte o Jogo tem de estar.

'(Float,Float)' representa as coordenadas em que as imagens começarão por ser apresentadas!

'[Picture]''s representam todas as listas de loads para poder apresentar as imagens quando necessário.

Float tem como efeito representar o número em que o tempo irá começar a ser contado.

Load de Imagens

loadMenus :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

loadJogar :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

loadNiveis :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

loadInstrucoes :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

loadVitoria :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

loadBackgrounds :: IO [Picture]

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, backgrounds, etc.

Janela

window :: Display

Neste caso, atribuimos um FullScreen, pois desejamos que o jogo seja aberto em tela cheia, seja qual for o monitor!

Frame Rate

fr :: Int

Atribuido para reagir ao tempo!

Estado Inicial

estadoInicial :: [Picture] -> [Picture] -> [Picture] -> [Picture] -> [Picture] -> [Picture] -> EstadoGloss

Representa como o jogo irá ser apresentado inicialmente, ou seja, ao ser aberto.

Desenhar Estado

draw :: EstadoGloss -> IO Picture

draw é a função que irá desenhar o jogo cada vez que muda alguma coisa, ou seja, reproduz algo para todos os estados possíveis.

O mesmo está organizado de acordo com cada "grupo" de estados possíveis, isto é, Tema, Menu, ModoJogo, Instruções, etc__.

Todas as funções aqui envolvidas estão num ficheiro onde se encontram todas as Auxiliares desta tarefa de modo a tornar o código mais fluído e legível.

Reage Evento

reageEvento :: Event -> EstadoGloss -> IO EstadoGloss

Esta função trata de atribuir funções às teclas e trazer-lhes consequências de acordo com o estado atual do jogo.

As mesmas estão organizadas de acordo com cada parte do jogo, menu, niveis, ModoJogo, etc.

Reage Tempo

Reage Tempo

reageTempo :: Float -> EstadoGloss -> IO EstadoGloss

Desta forma, todos os estados estão submetidos à mesma contagem de tempo. Isto é, quando o jogo é aberto, começa a ser contado o tempo e então, mesmo que o estado mude.

MainFunction

main :: IO ()

As funções loaded auxiliam o estadoInicial. Não há muito a dizer acerca desta função.