T5Grafica

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2021/22.

Documentation

data Tipos

Constructors

Nada

Select

Play

Level

Tips

Exit

Back

Lvl1

Lvl2

Lvl3

Lvl4

Lvl5

Lvl6

Lvl7

Lvl8

Lvl9

Lvl10

Texto

WinCima

WinBaixo

Continue

NiveisNormal

OpcoesNormal

GameNormal

BlockDude

DragonBall

AmongUs

Theme

Show1

Show2

Show3

Show4

Show5

Show6

Show7

Show8

Show9

Show10

Theme1

Theme2

Theme3

ATheme

DBTheme

NTheme

data Jogos

Constructors

Jogo1

Jogo2

Jogo3

Jogo4

Jogo5

Jogo6

Jogo7

Jogo8

Jogo9

Jogo10

data OpcoesPrincipal

Constructors

Jogar

Niveis

Continuar

Instrucoes

Tema

Sair

data OpcoesNiveis

Constructors

Nivel1

Nivel2

Nivel3

Nivel4

Nivel5

Nivel6

Nivel7

Nivel8

Nivel9

Nivel10

Voltar

data Menu

Constructors

Controlador OpcoesPrincipal Tipos

Selecionador OpcoesNiveis Tipos

ModoJogo Jogo Jogo Jogos Tipos

Pergaminho Tipos

ImagemVitoria OpcoesPrincipal Jogo Jogos Tipos

Temas Tipos Tipos

type EstadoGloss = (Menu, (Float, Float), [Picture], [Picture], [Picture], [Picture], [Picture], [Picture], Float)

Types

Nesta 'type':

- Menu vai representar em que parte o Jogo tem de estar.

- '(Float,Float)' representa as coordenadas em que as imagens começarão por ser apresentadas!

- '[Picture]''s representam todas as listas de loads para poder apresentar as imagens quando necessário.

- Float tem como efeito representar o número em que o tempo irá começar a ser contado.

loadMenus :: IO [Picture]

Load Imagens

Dividido em categorias, cada load está atribuido a uma parte do jogo, isto é, um load para o mapa, menu, etc.

loadJogar :: IO [Picture]

loadNiveis :: IO [Picture]

loadInstrucoes :: IO [Picture]

loadVitoria :: IO [Picture]

loadBackgrounds :: IO [Picture]

window :: Display

Janela

Neste caso, atribuimos um FullScreen, pois desejamos que o jogo seja aberto em tela cheia, seja qual for o monitor!

fr :: Int

FrameRate

Atribuido para reagir ao tempo!

estadoInicial :: [Picture] -> [Picture] -> [Picture] -> [Picture] -> [Picture] -> [Picture] -> EstadoGloss

Estado Inicial

draw :: EstadoGloss -> IO Picture

Desenha Estado

drawMenu :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

drawMenu2 :: ([Tipos], [Picture]) -> Int -> [Picture]

opcoes :: (Tipos, [Picture]) -> Picture

drawMapaN :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

drawMapa2N :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

drawPlayerN :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

drawMapaDB :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

drawMapa2DB :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

drawPlayerDB :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

drawMapaA :: (Mapa, [Picture]) -> (Int, Int) -> Float -> [Picture]

drawMapa2A :: ([Peca], [Picture]) -> Int -> Float -> [Picture]

drawPlayerA :: ([Peca], (Int, Int), [Picture]) -> Float -> [Picture]

pecasN :: (Peca, [Picture]) -> Float -> Picture

pecasA :: (Peca, [Picture]) -> Float -> Picture

pecasDB :: (Peca, [Picture]) -> Float -> Picture

drawNiveis :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

drawNiveis2 :: ([Tipos], [Picture]) -> Int -> [Picture]

levels :: (Tipos, [Picture]) -> Picture

drawInstrucoes :: ([Tipos], [Picture]) -> Int -> [Picture]

instruct :: (Tipos, [Picture]) -> Picture

drawVitoria :: ([[Tipos]], [Picture]) -> (Int, Int) -> [Picture]

drawVitoria2 :: ([Tipos], [Picture]) -> Int -> [Picture]

victory :: (Tipos, [Picture]) -> Picture

backgroundsDB :: (Tipos, [Picture]) -> Picture

backgroundsN :: (Tipos, [Picture]) -> Picture

backgroundsA :: (Tipos, [Picture]) -> Picture

porta :: Jogo -> (Int, Int)

findPorta :: Mapa -> Int -> Int -> (Int, Int)

reageEvento :: Event -> EstadoGloss -> IO EstadoGloss

Reage Evento

reageTempo :: Float -> EstadoGloss -> IO EstadoGloss

Reage Tempo

main :: IO ()

Main